requirements.txt
python-telegram-bot[ext] >=20.2
Flask>=2.3.2
pandas>=2.0.1
aiohttp>=3.8.4
